requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.24.4
openai==1.14.3
packaging==24.0
pgvector==0.2.5
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-core==2.16.3
Pygments==2.3.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pypdf2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==5.3.1
regex==2023.12.25
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
starlette==0.36.3
tqdm==4.66.2
typing-extensions==4.10.0
uvicorn==0.29.0
